Created the examples directory, containing components called "before" and "after"
The before directory is a component which illustrates typical issues that GCC has with the
Symbian source code. Compiling this component (sbs -b bld.inf -k -c armv5_udeb_gcce4_4_1)
will produce multiple errors, hence the use of "-k" to keep going.

The after directory is the same component with fixes applied - compiling in the after directory
should produce no errors. The aim is to keep the code very closely aligned to the before component,
so that they can be compared with visual diff tools such as Beyond Compare.

Sometimes it's valuable to be able to compile this code with the default production compiler,
(currently RVCT 2.2) so the MMP files will switch the name of the generated executable if a
different compiler is used.